+++ MacForge/MacForge/Views/Web/MF_featuredView.m
@@ -29,11 +29,17 @@ extern AppDelegate *myDelegate;
     }
 }
 
-- (void)drawRect:(NSRect)dirtyRect {
-    NSUInteger pad = self.frame.size.width - (4 * columns);
-    for (NSTableColumn* c in self.tv.tableColumns)
+- (void)adjustColumnWidth {
+    int multiplier = 4;
+    if (NSProcessInfo.processInfo.operatingSystemVersion.majorVersion == 11) multiplier = 25;
+    NSUInteger pad = self.frame.size.width - (multiplier * columns);
+    for (NSTableColumn* c in _tv.tableColumns)
         [c setWidth:pad/columns];
+}
+
+- (void)drawRect:(NSRect)dirtyRect {
     [self checkAndUpdate];
+    [self adjustColumnWidth];
 }
 
 - (void)updateColumCount {
@@ -42,22 +48,16 @@ extern AppDelegate *myDelegate;
     if (_tv.numberOfColumns > columns)
         for (int i = 0; i < nuke; i++)
             [_tv removeTableColumn:_tv.tableColumns.lastObject];
-    
+
     // add needed columns
     long give = columns - _tv.numberOfColumns;
     if (_tv.numberOfColumns < columns) {
         for (int i = 0; i < give; i++) {
             NSString *identify = [NSString stringWithFormat:@"Col%d", (int)_tv.numberOfColumns + 1];
-            NSTableColumn * column = [[NSTableColumn alloc] initWithIdentifier:identify];
-            [column setWidth:self.frame.size.width/columns];
-            [_tv addTableColumn:column];
+            [_tv addTableColumn:[NSTableColumn.alloc initWithIdentifier:identify]];
         }
     }
-    
-    // set columns to equal width
-    for (NSTableColumn *col in _tv.tableColumns)
-        [col setWidth:self.frame.size.width/columns];
-    
+
     // redraw and fit
     [_tv reloadData];
 }
@@ -69,23 +69,27 @@ extern AppDelegate *myDelegate;
         smallArray = NSMutableArray.new;
         
         // Create a table view
-        _tv = [[NSTableView alloc] initWithFrame:NSMakeRect(0, 0, 500, 500)];
+        
+        NSRect theFrame = self.frame;
+        if (MF_extra.sharedInstance.macOS >= 16) theFrame.size.height += 38;
+        
+        _tv = [NSTableView.alloc initWithFrame:theFrame];
         _tv.delegate = self;
         _tv.dataSource = self;
         _tv.gridColor = NSColor.clearColor;
         _tv.backgroundColor = NSColor.clearColor;
         _tv.headerView = nil;
         _tv.selectionHighlightStyle = NSTableViewSelectionHighlightStyleNone;
-        
+                
         // Create a scroll view and embed the table view in the scroll view, and add the scroll view to our window.
-        NSScrollView * tableContainer = [[NSScrollView alloc] initWithFrame:self.frame];
-        tableContainer.documentView = _tv;
-        tableContainer.drawsBackground = false;
-        tableContainer.hasVerticalScroller = true;
-        tableContainer.hasHorizontalScroller = false;
-        tableContainer.horizontalScrollElasticity = NSScrollElasticityNone;
-        tableContainer.autoresizingMask = NSViewWidthSizable | NSViewHeightSizable;
-        [self addSubview:tableContainer];
+        _sv = [NSScrollView.alloc initWithFrame:theFrame];
+        _sv.documentView = _tv;
+        _sv.drawsBackground = false;
+        _sv.hasVerticalScroller = true;
+        _sv.hasHorizontalScroller = false;
+        _sv.horizontalScrollElasticity = NSScrollElasticityNone;
+        _sv.autoresizingMask = NSViewWidthSizable | NSViewHeightSizable;
+        [self addSubview:_sv];
         
         dispatch_queue_t backgroundQueue = dispatch_queue_create("com.macenhance.MacForge", 0);
         dispatch_async(backgroundQueue, ^{
@@ -101,11 +105,12 @@ extern AppDelegate *myDelegate;
         for (int i = 0; i < columns; i++) {
             NSString *identify = [NSString stringWithFormat:@"Col%d", i];
             NSTableColumn * column = [[NSTableColumn alloc] initWithIdentifier:identify];
+            [column setResizingMask:NSTableColumnNoResizing];
             [column setWidth:self.frame.size.width/columns];
             [_tv addTableColumn:column];
         }
     });
-
+    
     [self checkAndUpdate];
 }