+++ MacForge/Pods/Target Support Files/FirebaseCoreDiagnostics/FirebaseCoreDiagnostics-umbrella.h
@@ -11,6 +11,13 @@
 #endif
 
 #import "firebasecore.nanopb.h"
+#import "GoogleDataTransportInternal.h"
+#import "GULAppEnvironmentUtil.h"
+#import "GULHeartbeatDateStorage.h"
+#import "GULKeychainStorage.h"
+#import "GULKeychainUtils.h"
+#import "GULSecureCoding.h"
+#import "GULLogger.h"
 #import "FIRCoreDiagnosticsData.h"
 #import "FIRCoreDiagnosticsInterop.h"