+++ MacForge/Pods/Target Support Files/Pods-purchaseValidationApp/Pods-purchaseValidationApp-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_Pods_purchaseValidationApp : NSObject
-@end
-@implementation PodsDummy_Pods_purchaseValidationApp
-@end